Refactor bot structure and command handling
* Split commands from index.ts into src/commands/utility directory using a dynamic loader
* Implement a standardized Command interface
* Introduce MyClient which extends Client with a Commands Collection
* Set up dynamic SlashCommand deployer script (deploy-commands.ts)
* Move generic functions to src/utils/helpers.ts and update early returns
* Setup and replace console.log with pino and pino-pretty logger throughout the app
* Added package scripts for new deploy and lint flows
