+++ package.json
@@ -12,7 +12,8 @@
     "lint": "eslint src",
     "lint:fix": "eslint src --fix",
     "format": "prettier --write 'src/**/*.{ts,json,md}'",
-    "test": "jest"
+    "test": "jest",
+    "deploy": "tsx src/deploy-commands.ts"
   },
   "keywords": [],
   "author": "",
@@ -27,6 +28,7 @@
     "joi": "^18.1.2",
     "moment": "^2.30.1",
     "pino": "^10.3.1",
+    "pino-pretty": "^13.1.3",
     "rimraf": "^6.1.3"
   },
   "devDependencies": {
