+++ src/types/command.ts
@@ -1,0 +1,6 @@
+import { ChatInputCommandInteraction, SlashCommandBuilder } from 'discord.js';
+
+export interface Command {
+  data: ReturnType<SlashCommandBuilder['toJSON']>
+  execute: (_interaction: ChatInputCommandInteraction) => Promise<void>
+}
